Moving MetricsDataModel to StatsDataModel to prepare for event driven metrics capture
