+++ src/main/scala/zio/zmx/metrics/package.scala
@@ -17,7 +17,7 @@ package object metrics {
 
     def increment(name: String) = ZIO.accessM[ZMetrics](_.get.increment(name))
 
-    def count[R, E, A](name: String)(e: ZIO[R, E, A]): ZIO[R with ZMetrics, Any, A] = for {
+    def count[R, E, A](name: String)(e: ZIO[R, E, A]): ZIO[R with ZMetrics, E, A] = for {
       r <- e
       _ <- increment(name)
     } yield r
@@ -30,4 +30,11 @@ package object metrics {
     ZLayer.succeed(new StatsdInstrumentation(cfg))
 
   lazy val empty: ZLayer[Any, Nothing, ZMetrics] = ZLayer.succeed(new EmptyInstrumentation())
+
+  implicit class MZio[R, E, A](z: ZIO[R, E, A]) {
+
+    import ZMetrics._
+
+    def counted(name: String): ZIO[R with ZMetrics, E, A] = count[R, E, A](name)(z)
+  }
 }
