+++ src/test/scala/zio/zmx/InstrumentedSample.scala
@@ -6,7 +6,8 @@ import zio.zmx.metrics._
 trait InstrumentedSample {
 
   def doSomething    = ZMetrics.count("myCounter")(ZIO.succeed(print(".")))
-  def countSomething = ZIO.foreach_(1.to(100))(_ => doSomething)
+  def doSomething2   = ZIO.succeed(print(".")).counted("myCounter2")
+  def countSomething = ZIO.foreach_(1.to(100))(_ => doSomething2.zip(doSomething))
 
   def program: ZIO[ZEnv with ZMetrics, Nothing, ExitCode] = for {
     _ <- countSomething.absorbWith(_ => new Exception("Boom")).orDie
